refactor: #232 퀘스트 조회 헤더 뷰 상속 구조로 변경
